requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Serverless adapter for Vercel
mangum>=0.17.0

# Environment variables
python-dotenv>=1.0.0

# Core dependencies for RAG functionality
pydantic>=2.5.0
pydantic-settings>=2.1.0
typing-extensions>=4.8.0

# OpenAI SDK
openai>=1.3.5

# Google Gemini SDK (FREE embeddings and chat)
google-generativeai>=0.4.1

# Vector database
qdrant-client>=1.7.0

# Database adapter
psycopg2-binary>=2.9.9

# Rate limiting
slowapi>=0.1.9

# HTTP client for testing
httpx>=0.25.0

# Markdown processing
markdown>=3.5.0

# Token counting (OpenAI tokenizer)
tiktoken>=0.5.0

# Additional utilities
python-multipart>=0.0.6